requirements.txt
absl-py==0.11.0
lit-nlp==0.3
numpy>=1.19.0
pandas>=1.0.4
torch==1.9.0
transformers==4.10.2
